Fix a pb for inclusion of ROS & python interpreter in ART OpenRTM.
Pb: The openrtm architecture loads a shared library to implement the services of
an OpenRTM node. The scheduling properties are ensured by the operating system
running openrtm. Then on the robot, when the shared library is loaded it is
run inside the real-time loop. As sot-hrp-2 was creating the interpreter and
the ros-node in the constructor this used to break HRP-2 real-time OS.

This commit change the initialization of the ros-node to use a multi-threaded
spinner. Then ros calls are handled in a seperate non-realtime thread.