[fix] Geometric simulation
Parsing the arguments given to the SoT implies an access to the
list of robots inside the SoT (due to a change in the ParameterServer in
sot-core), which is empty at the moment.
Initializing the ROS node creates the robot inside the SoT, without meaningful
side-effects.

Therefore, this simple fix swaps those two steps.